refactor: split tech_extraction and cache unlock maps (#4412)
Stop rebuilding regiment/ship unlock maps on every access and split the mixed-concern tech_extraction module so catalog query, extraction caps, and unlock helpers can evolve independently.
