+++ SPEC/game/tech-tree-gathering.md
@@ -6,7 +6,7 @@
 
 ## Catalog and implementation
 
-The **gathering table below is the GDD source of truth** for ids, effects, and prerequisites. Prerequisites may reference techs defined in **other category docs** (e.g. [tech-tree-labour-economy.md](tech-tree-labour-economy.md) for `university`, [tech-tree-transport.md](tech-tree-transport.md) for `dynamite`, [tech-tree-new-world.md](tech-tree-new-world.md) for `precious_stone_mining` / `precious_metals_mining`, [tech-tree-military.md](tech-tree-military.md) for artillery-related techs). The program-level catalog lives in code (`colonizethis_data` tech modules) and uses these same ids. **Unlock and prerequisite validation** use the shared tech catalog. **Extraction caps** are resolved per resource per [tech-and-extraction-cap.md](tech-and-extraction-cap.md) (`tech_extraction.dart`). Tech table source is program-level constants; future ruleset-driven catalog per [ruleset-config.md](../program/ruleset-config.md).
+The **gathering table below is the GDD source of truth** for ids, effects, and prerequisites. Prerequisites may reference techs defined in **other category docs** (e.g. [tech-tree-labour-economy.md](tech-tree-labour-economy.md) for `university`, [tech-tree-transport.md](tech-tree-transport.md) for `dynamite`, [tech-tree-new-world.md](tech-tree-new-world.md) for `precious_stone_mining` / `precious_metals_mining`, [tech-tree-military.md](tech-tree-military.md) for artillery-related techs). The program-level catalog lives in code (`colonizethis_data` tech modules) and uses these same ids. **Unlock and prerequisite validation** use the shared tech catalog. **Extraction caps** are resolved per resource per [tech-and-extraction-cap.md](tech-and-extraction-cap.md) (`tech_extraction_caps.dart`). Tech table source is program-level constants; future ruleset-driven catalog per [ruleset-config.md](../program/ruleset-config.md).
 
 ---
